Don't skip all buffers in diff mode
The s:_skip_file function is generic enough to opt-out syntax checking
in diff buffers if the user wants to by setting b:syntastic_skip_checks,
but there is no opt-in otherwise.

The problem is that this breaks some workflows, so it can't be included
in the main branch. See the discussions:

https://github.com/scrooloose/syntastic/issues/467
https://github.com/scrooloose/syntastic/pull/1184

If you don't use vimdiff or g:syntastic_check_on_open you should be safe
with this patch.